Merging dev to main

+++ lib/actions.ts
@@ -3,7 +3,8 @@
 import { revalidatePath } from "next/cache";
 import { redirect } from "next/navigation";
 import { API_URL } from "./config";
-import { Product, ProductsResponse } from "@/app/types";
+import { ProductsResponse } from "@/app/types";
+import { GetSession } from "./authActions";
 
 export const FetchAllCategories = async () => {
   const data = await fetch(`${API_URL}/categories`).then((res) => res.json());
@@ -12,7 +13,7 @@ export const FetchAllCategories = async () => {
 };
 
 export const FetchAllProducts = async (
-  limit = 6 ,
+  limit = 6,
   page = 1,
   sort = "id",
   order = "asc",
@@ -27,6 +28,12 @@ export const FetchAllProducts = async (
     title_like: query,
     _expand: "category",
   });
+
+  const session = GetSession();
+
+  if (!session) {
+    throw new Error("Unauthorized");
+  }
 
   if (categoryId) {
     params.set("categoryId", categoryId);
@@ -44,6 +51,11 @@ export const FetchAllProducts = async (
 };
 
 export const FetchSingleProductById = async (id: string) => {
+  const session = GetSession();
+
+  if (!session) {
+    throw new Error("Unauthorized");
+  }
   const data = await fetch(`${API_URL}/products/${id}`).then((res) =>
     res.json(),
   );
@@ -51,16 +63,12 @@ export const FetchSingleProductById = async (id: string) => {
   return data;
 };
 
-/*  
-    "title",
-    "price",
-    "description",
-    "thumbnail",
-    "categoryId",
-    "brand",
-*/
-
 export const CreateProduct = async (formData: FormData) => {
+  const session = GetSession();
+
+  if (!session) {
+    throw new Error("Unauthorized");
+  }
   const title = formData.get("title") as string;
   const price = formData.get("price") as string;
   const description = formData.get("description") as string;
@@ -68,6 +76,18 @@ export const CreateProduct = async (formData: FormData) => {
   const categoryId = formData.get("categoryId") as string;
   const stock = formData.get("stock") as string;
   const brand = formData.get("brand") as string;
+
+  let availabilityStatus: string;
+
+  if (Number(stock) >= 25) {
+    availabilityStatus = "In Stock";
+  } else if (Number(stock) < 25 && Number(stock) > 0) {
+    availabilityStatus = "Low Stock";
+  } else if (Number(stock) === 0) {
+    availabilityStatus = "Out of Stock";
+  } else {
+    throw new Error();
+  }
 
   const newProduct = {
     title,
@@ -77,16 +97,30 @@ export const CreateProduct = async (formData: FormData) => {
     price: parseInt(price, 10),
     categoryId: parseInt(categoryId, 10),
     stock: parseInt(stock, 10),
+    availabilityStatus,
   };
+
+  console.log(newProduct);
 
   const res = await fetch(`${API_URL}/products`, {
     method: "POST",
     headers: { "Content-Type": "application/json" },
     body: JSON.stringify(newProduct),
   });
+
+  if (!res.ok) {
+    const error = await res.text();
+    console.error(error);
+    throw new Error("Failed to create product");
+  }
 };
 
 export const UpdateProduct = async (formData: FormData) => {
+  const session = GetSession();
+
+  if (!session) {
+    throw new Error("Unauthorized");
+  }
   const id = formData.get("id") as string;
   const title = formData.get("title") as string;
   const price = formData.get("price") as string;
@@ -96,6 +130,18 @@ export const UpdateProduct = async (formData: FormData) => {
   const stock = formData.get("stock") as string;
   const brand = formData.get("brand") as string;
 
+  let availabilityStatus: string;
+
+  if (Number(stock) >= 25) {
+    availabilityStatus = "In Stock";
+  } else if (Number(stock) < 25 && Number(stock) > 0) {
+    availabilityStatus = "Low Stock";
+  } else if (Number(stock) === 0) {
+    availabilityStatus = "Out of Stock";
+  } else {
+    throw new Error();
+  }
+
   const newProduct = {
     title,
     brand,
@@ -104,6 +150,7 @@ export const UpdateProduct = async (formData: FormData) => {
     price: parseInt(price, 10),
     categoryId: parseInt(categoryId, 10),
     stock: parseInt(stock, 10),
+    availabilityStatus,
   };
 
   const res = await fetch(`${API_URL}/products/${id}`, {
@@ -114,11 +161,17 @@ export const UpdateProduct = async (formData: FormData) => {
 
   const data = await res.json();
 
-  revalidatePath("/");
+  revalidatePath("/products");
   redirect("/");
 };
 
 export async function UpdateProductBind(id: string, formData: FormData) {
+  const session = GetSession();
+
+  if (!session) {
+    throw new Error("Unauthorized");
+  }
+
   const title = formData.get("title") as string;
   const price = formData.get("price") as string;
   const description = formData.get("description") as string;
@@ -144,13 +197,17 @@ export async function UpdateProductBind(id: string, formData: FormData) {
   });
 
   const data = await res.json();
-  console.log(data);
 
   revalidatePath("/");
   redirect("/");
 }
 
 export async function deleteProduct(formData: FormData) {
+  const session = GetSession();
+
+  if (!session) {
+    throw new Error("Unauthorized");
+  }
   const id = formData.get("id") as string;
 
   const res = await fetch(`${API_URL}/products/${id}`, {
@@ -164,6 +221,11 @@ export async function deleteProduct(formData: FormData) {
 }
 
 export const DeleteProductById = async (id: string) => {
+  const session = GetSession();
+
+  if (!session) {
+    throw new Error("Unauthorized");
+  }
   const res = await fetch(`${API_URL}/products/${id}`, {
     method: "DELETE",
     headers: { "Content-Type": "application/json" },
@@ -177,6 +239,11 @@ export const DeleteProductById = async (id: string) => {
 };
 
 export async function deleteProductBind(id: string) {
+  const session = GetSession();
+
+  if (!session) {
+    throw new Error("Unauthorized");
+  }
   const res = await fetch(`${API_URL}/products/${id}`, {
     method: "DELETE",
   });
